rename and remove files
