--- index2.js
@@ -1,3 +1,0 @@
-const weh = require('./lib/weh2')
-
-module.exports = weh
