+++ lib/config.js
@@ -21,12 +21,7 @@ function generateConfig (conf = {}) {
   return result
 }
 
-function mergeConfigs (conf1, conf2) {
-  return merge(conf1, conf2)
-}
-
 module.exports = {
   DEFAULT_CONFIG,
-  generateConfig,
-  mergeConfigs
+  generateConfig
 }
