+++ lib/hooks.js
@@ -1,33 +1,33 @@
-const ware = require('ware')
+const trough = require('trough')
 
-class Hooks {
-  constructor () {
-    const names = [
-      'pre_read',
-      'post_read',
-      'pre_write',
-      'post_write'
-    ]
+function hooks () {
+  const names = [
+    'pre_read',
+    'post_read',
+    'pre_write',
+    'post_write'
+  ]
 
-    this.registry = {}
+  const registry = {}
 
-    names.forEach(name => {
-      this.registry[name] = ware()
-    })
-  }
+  names.forEach(function (name) {
+    registry[name] = trough()
+  })
 
-  register (name, fn) {
-    this.registry[name].use(fn)
-  }
+  return {
+    register (name, fn) {
+      registry[name].use(fn)
+    },
 
-  run (name, ...args) {
-    return new Promise((resolve, reject) => {
-      this.registry[name].run(...args, (err, config) => {
-        if (err) reject(err)
-        resolve(config)
+    run (name, ...args) {
+      return new Promise((resolve, reject) => {
+        registry[name].run(...args, (err, result) => {
+          if (err) reject(err)
+          resolve(result)
+        })
       })
-    })
+    }
   }
 }
 
-module.exports = Hooks
+module.exports = hooks
